change int -> double

+++ src/MainModules/AbstractClasses/CalculateAbs.java
@@ -3,7 +3,7 @@ package MainModules.AbstractClasses;
 import DataPackage.Variable;
 
 public abstract class CalculateAbs extends Variable {
-   public abstract int sum(int x, int y);
+   public abstract double sum(double x, double y);
 
 
 }
